+++ hardware/firmware/audio_board/helpers/generate_graph.py
@@ -26,7 +26,8 @@ outputs = [
 ]
 print("// clang-format off")
 print()
-print('#include "channel_strip.h"')
+print('#include "channel_input.h"')
+print('#include "channel_output.h"')
 print('#include <Audio.h>')
 print("// GUItool: begin automatically generated code")
