+++ hardware/firmware/audio_board/src/channel_info.h
